third_party/blink/web_tests/platform/win/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/parallel-dispatch.https.any.worker_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] dispatch queues behind readBuffer
  assert_array_equals: expected property 0 to be 2 but got 20 (expected array [2, 20] got [20, 20])
Harness: the test ran to completion.
